pkg/oniontree/monitor: Watch the filesystem changes
Previously, monitor had a timer which was used to trigger a rescan of
OnionTree directory. The timer had a bug which effectively disabled the
rescaning.

Instead of fixing the issue by changing how the timer is initialized,
this beening the root cause, I have changed the design to use filesystem
notifications instead. This had been on my TODO list for some time.

For notifications, fsnotify is used. The monitor watches events on two
OnionTree directories: tagged/dead and unsorted. When a new event is
received, it is forwarded to eventCh which then triggers the rescan.