services/vworp: Do not check bool returned by GetOnlineLinks
When GetOnlineLinks returns, it returns a slice populated with online
links and a boolean value that is set to false when service ID (used as
a key) is not present in the underlaying map.

When vworp starts, first few seconds or so there might be no data in the
map, if user clicks on redirect page, they might see 404 Not Found error
accompanied by the message that the service does not exist.

This commit disables the check entirely and the error 404 Not Found is
replaced by 200 OK with the message that the service is offline.