services/vworp: Improve error handling
This change is two-fold. First, it's a major refactoring of the previous
code. Error handling is done inside a dedicated handleOops handler which
is called by other handlers and is also registered as a standalone URL
handler.

Second, oops id is translated to HTTP status code and sent back to the
client. This improves log readability. Before, all the responses were
hidden under `302 See Other` status code.